test: move obsolete-snapshot tests to a clean sibling file
snapshots/snapshot.test.ts has an env-dependent pre-existing failure
(error snapshots needs FORCE_COLOR) that would mask the fail-before
check; keep the new coverage in its own file.
